Add calculator logic in script file and refactored
The logic has variables storing the first/second operand and the
operator, with values undefined if none has been input. Each time a
button is clicked, a block of code (corresponding to which of these
variables are currently defined or not and whether the first operand
has been finalized or not) is ran. I refactored some repeated code in
the script creation of certain html elements.